offline-phase: Prove & verify ciphertext PoK in key exchange
When the two parties setup a global MAC key they prove knowledge of the
key to avoid adversarial waiting that may allow a party to induce
algebraic structure on the key by choosing their key appropriately.

This PoK also asserts that the noise is within a given bound, i.e. that
the ciphertext may be properly decrypted.